Updated ics file with GEO
Have written an addition to the code that looks up the city, country
and gets lat/lon and adds that to each event if it resolves.

I have a PR on the underlying Data::ICal module that resolves an
issue where it is escaping the semi-colon.

Will commit the code once that is accepted as I need a fresh cpanfile